feat(#1306): remove ability to turn off partitioning and optimising
* got rid of providers and unnecessary interface and implementation of
  optimiser and partitioner
