--- generator/src/main/java/com/scottlogic/deg/generator/decisiontree/treepartitioning/NoopTreePartitioner.java
@@ -1,28 +1,0 @@
-/*
- * Copyright 2019 Scott Logic Ltd
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.scottlogic.deg.generator.decisiontree.treepartitioning;
-
-import com.scottlogic.deg.generator.decisiontree.DecisionTree;
-
-import java.util.stream.Stream;
-
-public class NoopTreePartitioner implements TreePartitioner {
-    @Override
-    public Stream<DecisionTree> splitTreeIntoPartitions(DecisionTree decisionTree) {
-        return Stream.of(decisionTree);
-    }
-}
