--- generator/src/main/java/com/scottlogic/deg/generator/guice/DecisionTreeOptimiserProvider.java
@@ -1,31 +1,0 @@
-/*
- * Copyright 2019 Scott Logic Ltd
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.scottlogic.deg.generator.guice;
-
-import com.google.inject.Provider;
-import com.scottlogic.deg.generator.decisiontree.DecisionTreeOptimiser;
-import com.scottlogic.deg.generator.decisiontree.MostProlificConstraintOptimiser;
-
-public class DecisionTreeOptimiserProvider implements Provider<DecisionTreeOptimiser> {
-    public DecisionTreeOptimiserProvider() {
-    }
-
-    @Override
-    public DecisionTreeOptimiser get() {
-        return new MostProlificConstraintOptimiser();
-    }
-}
